requirements.txt
nltk==3.8.1
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2
pandas==2.1.1
numpy==1.26.2
python-dateutil==2.8.2
newspaper3k==0.2.8
beautifulsoup4==4.12.2
requests==2.31.0
google-generativeai==0.3.1
python-dotenv==1.0.0
flask==2.3.3 